chore: typescript support
- typecheck
- lint
- build
- fix things that updating eslint found
- bump linting to node 16 because eslint likes it?
- binding-mock's exports are changed
