+++ .eslintignore
@@ -1,2 +1,3 @@
 packages/*/node_modules
 docs
+packages/*/dist
