+++ .github/workflows/test-merge.yml
@@ -15,9 +15,8 @@ jobs:
     - uses: actions/checkout@v2
     - uses: actions/setup-node@v2
       with:
-        node-version: 14
+        node-version: 16
     - run: npm ci
-    - run: npm run rebuild
     - run: npm test
     - run: npm run coverage
       env:
