+++ .github/workflows/test.yml
@@ -8,9 +8,10 @@ jobs:
     - uses: actions/checkout@v2
     - uses: actions/setup-node@v2
       with:
-        node-version: 14
+        node-version: 16
     - run: npm ci
     - run: npm run lint
+    - run: npm run typecheck
   test:
     runs-on: ${{ matrix.config.os }}
     strategy:
@@ -27,7 +28,6 @@ jobs:
       with:
         node-version: ${{ matrix.node-version }}
     - run: npm ci
-    - run: npm run rebuild
     - run: npm test
     - run: npm run coverage
       env:
