+++ package.json
@@ -8,41 +8,53 @@
   },
   "license": "MIT",
   "scripts": {
-    "rebuild": "lerna run --stream rebuild",
+    "build": "lerna run build",
     "coverage": "nyc report --reporter=text-lcov > coverage.lcov && codecov",
     "format": "eslint packages test --fix",
     "generate": "plop",
-    "lint": "eslint packages test && lerna run --stream lint",
+    "lint": "eslint packages test",
     "outdated": "lerna exec --no-bail npm outdated && npm outdated",
     "postinstall": "lerna bootstrap --no-ci",
     "publish": "lerna publish --exact",
-    "test": "nyc --reporter=html --reporter=text --reporter lcovonly mocha",
     "test:arduino": "TEST_PORT=$(./bin/find-arduino.js) npm test",
-    "test:watch": "mocha -w"
+    "test:watch": "mocha -w",
+    "test": "nyc --reporter=html --reporter=text --reporter lcovonly mocha",
+    "typecheck": "tsc"
   },
   "repository": {
     "type": "git",
     "url": "git://github.com/serialport/node-serialport.git"
   },
   "devDependencies": {
-    "chai-subset": "^1.6.0",
+    "@tsconfig/node12": "^1.0.9",
+    "@types/chai": "^4.3.0",
+    "@types/chai-subset": "^1.3.3",
+    "@types/debug": "^4.1.7",
+    "@types/mocha": "^9.1.0",
+    "@types/sinon": "^10.0.8",
+    "@typescript-eslint/eslint-plugin": "^5.10.0",
+    "@typescript-eslint/parser": "^5.10.0",
     "chai": "^4.3.4",
+    "chai-subset": "^1.6.0",
     "codecov": "^3.8.3",
+    "esbuild": "^0.14.13",
+    "esbuild-register": "^3.3.2",
+    "eslint": "^8.2.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.25.3",
     "eslint-plugin-mocha": "^9.0.0",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-promise": "^6.0.0",
-    "eslint": "^8.2.0",
-    "lerna-changelog": "^2.2.0",
     "lerna": "^4.0.0",
+    "lerna-changelog": "^2.2.0",
     "mocha": "^9.1.3",
     "nyc": "^15.1.0",
     "plop": "^2.7.6",
     "prettier": "^2.4.1",
     "proxyquire": "^2.1.3",
-    "sinon": "^12.0.1"
+    "sinon": "^12.0.1",
+    "typescript": "^4.5.5"
   },
   "funding": "https://opencollective.com/serialport/donate",
   "changelog": {
