+++ packages/binding-mock/.npmignore
@@ -1,3 +1,6 @@
 .DS_Store
 *.test.js
 CHANGELOG.md
+lib
+tsconfig.json
+tsconfig-build.json
