+++ packages/binding-mock/README.md
@@ -1,11 +1,11 @@
-# @serialport/BindingMock
+# @serialport/binding-mock
 
-This stream does some neat stuff.
+This does some neat stuff.
 
 This is why you'd use it.
 
 This is how you use it.
 ```js
-const bindingMock = new BindingMock()
-
+import { MockBinding } from '@serialport/binding-mock'
+const MockBinding = new MockBinding()
 ```
