+++ packages/binding-mock/tsconfig-build.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../../tsconfig-build.json",
+  "compilerOptions": {
+    "outDir": "dist",
+  },
+  "exclude": [
+    "node_modules",
+    "**/*.test.ts",
+  ]
+}
