+++ packages/stream/lib/index.test.js
@@ -5,7 +5,7 @@ chai.use(require('chai-subset'))
 const assert = chai.assert
 
 const SerialPort = require('../')
-const MockBinding = require('@serialport/binding-mock')
+const { MockBinding } = require('@serialport/binding-mock')
 
 describe('SerialPort', () => {
   let sandbox
