--- test/initializers/assert.js
@@ -1,4 +1,0 @@
-const chai = require('chai')
-const chaiSubset = require('chai-subset')
-global.assert = chai.assert
-chai.use(chaiSubset)
