+++ tsconfig-build.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "declaration": true,
+    "declarationMap": false,
+    "noEmitOnError": true,
+    "noEmit": false
+  }
+}
