+++ tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "extends": "@tsconfig/node12/tsconfig.json",
+  "compilerOptions": {
+    "baseUrl": ".",
+    "paths": {
+      "@serialport/*": ["packages/*/lib"],
+      "serialport": ["packages/serialport/lib"]
+    },
+    "allowJs": true,
+    "checkJs": false,
+    "noEmit": true
+  }
+}
